requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
appdirs==1.4.4
astor==0.8.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
black==21.5b2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
execnet==1.9.0
idna==3.2
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.3.0
regex==2021.4.4
requests==2.26.0
selenium==3.141.0
six==1.16.0
texttable==1.6.4
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.2
